endPoints/requirements.txt
fastapi>=0.111.0
uvicorn>=0.29.0
langchain-core>=0.2.0
langchain-google-genai>=0.1.0
python-dotenv>=1.0.1
pydantic>=2.7.3
